app/requirements.txt
boto3
prometheus_client
pyYAML
python-dateutil
